split FinalGamePlayStats into two
