+++ GW2EIEvtcParser/EIData/Statistics/FinalGameplayStats.cs
@@ -1,113 +1,123 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using GW2EIEvtcParser.ParsedData;
+using static GW2EIEvtcParser.EIData.Buff;
 
 namespace GW2EIEvtcParser.EIData
 {
     public class FinalGameplayStats
     {
-        public int TotalDamageCount { get; }
-        public int DirectDamageCount { get; }
-        public int ConnectedDamageCount { get; }
-        public int ConnectedDirectDamageCount { get; }
-        public int CritableDirectDamageCount { get; }
-        public int CriticalCount { get; }
-        public int CriticalDmg { get; }
-        public int FlankingCount { get; }
-        public int GlanceCount { get; }
-        public int AgainstMovingCount { get; }
-        public int Missed { get; }
-        public int Blocked { get; }
-        public int Evaded { get; }
-        public int Interrupts { get; }
-        public int Invulned { get; }
-        public int Killed { get; }
-        public int Downed { get; }
+        // Rates
+        public int Wasted { get; }
+        public double TimeWasted { get; }
+        public int Saved { get; }
+        public double TimeSaved { get; }
+        public double StackDist { get; }
+        public double DistToCom { get; }
 
+        // boons
+        public double AvgBoons { get; }
+        public double AvgActiveBoons { get; }
+        public double AvgConditions { get; }
+        public double AvgActiveConditions { get; }
 
-        internal FinalGameplayStats(ParsedEvtcLog log, long start, long end, AbstractSingleActor actor, AbstractSingleActor target)
+        // Counts
+        public int SwapCount { get; }
+
+        public double SkillCastUptime { get; }
+        public double SkillCastUptimeNoAA { get; }
+
+        private static double GetDistanceToTarget(AbstractSingleActor actor, ParsedEvtcLog log, long start, long end, IReadOnlyList<Point3D> reference)
         {
-            IReadOnlyList<AbstractHealthDamageEvent> dls = actor.GetDamageEvents(target, log, start, end);
-            foreach (AbstractHealthDamageEvent dl in dls)
+            var positions = actor.GetCombatReplayPolledPositions(log).Where(x => x.Time >= start && x.Time <= end).ToList();
+            int offset = actor.GetCombatReplayPolledPositions(log).Count(x => x.Time < start);
+            if (positions.Count > 1 && reference.Count > 0)
             {
-                if (dl.From == actor.AgentItem)
+                var distances = new List<float>();
+                for (int time = 0; time < positions.Count; time++)
                 {
-                    if (!(dl is NonDirectHealthDamageEvent))
-                    {
-                        if (dl.HasHit)
-                        {
-                            if (SkillItem.CanCrit(dl.SkillId, log.LogData.GW2Build))
-                            {
-                                if (dl.HasCrit)
-                                {
-                                    CriticalCount++;
-                                    CriticalDmg += dl.HealthDamage;
-                                }
-                                CritableDirectDamageCount++;
-                            }
-                            if (dl.IsFlanking)
-                            {
-                                FlankingCount++;
-                            }
 
-                            if (dl.HasGlanced)
-                            {
-                                GlanceCount++;
-                            }
-                            ConnectedDirectDamageCount++;
-                        }
+                    float deltaX = positions[time].X - reference[time + offset].X;
+                    float deltaY = positions[time].Y - reference[time + offset].Y;
+                    //float deltaZ = positions[time].Z - StackCenterPositions[time].Z;
 
-                        if (dl.IsBlind)
-                        {
-                            Missed++;
-                        }
-                        if (dl.IsEvaded)
-                        {
-                            Evaded++;
-                        }
-                        if (dl.IsBlocked)
-                        {
-                            Blocked++;
-                        }
-                        if (!dl.DoubleProcHit)
-                        {
-                            DirectDamageCount++;
-                        }
-                    }
-                    if (dl.IsAbsorbed)
-                    {
-                        Invulned++;
-                    }
-                    if (!dl.DoubleProcHit)
-                    {
-                        TotalDamageCount++;
-                    }
 
-                    if (dl.HasHit)
-                    {
-                        ConnectedDamageCount++;
-                        if (dl.AgainstMoving)
-                        {
-                            AgainstMovingCount++;
-                        }
-                    }
+                    distances.Add((float)Math.Sqrt(deltaX * deltaX + deltaY * deltaY));
                 }
+                return distances.Sum() / distances.Count;
+            }
+            else
+            {
+                return -1;
+            }
+        }
 
-                if (!(dl is NonDirectHealthDamageEvent))
+        internal FinalGameplayStats(ParsedEvtcLog log, long start, long end, AbstractSingleActor actor)
+        {
+            // If dummy actor, stop
+            if (actor.IsFakeActor)
+            {
+                return;
+            }
+            long duration = end - start;
+            foreach (AbstractCastEvent cl in actor.GetCastEvents(log, start, end))
+            {
+                switch (cl.Status)
                 {
-                    if (dl.HasInterrupted)
-                    {
-                        Interrupts++;
-                    }
+                    case AbstractCastEvent.AnimationStatus.Interrupted:
+                        Wasted++;
+                        TimeWasted += cl.SavedDuration;
+                        break;
+                    case AbstractCastEvent.AnimationStatus.Reduced:
+                        Saved++;
+                        TimeSaved += cl.SavedDuration;
+                        break;
                 }
-                if (dl.HasKilled)
+                if (cl.Skill.IsSwap)
                 {
-                    Killed++;
+                    SwapCount++;
                 }
-                if (dl.HasDowned)
+            }
+            TimeSaved = Math.Round(TimeSaved / 1000.0, ParserHelper.TimeDigit);
+            TimeWasted = -Math.Round(TimeWasted / 1000.0, ParserHelper.TimeDigit);
+            //
+            foreach (AbstractCastEvent cl in actor.GetIntersectingCastEvents(log, start, end))
+            {
+                var value = Math.Min(cl.EndTime, end) - Math.Max(cl.Time, start);
+                SkillCastUptime += value;
+                if (!cl.Skill.AA)
                 {
-                    Downed++;
+                    SkillCastUptimeNoAA += value;
                 }
-
+            }
+            long timeInCombat = Math.Max(actor.GetTimeSpentInCombat(log, start, end), 1);
+            SkillCastUptime /= timeInCombat;
+            SkillCastUptimeNoAA /= timeInCombat;
+            SkillCastUptime = Math.Round(100.0 * SkillCastUptime, ParserHelper.TimeDigit);
+            SkillCastUptimeNoAA = Math.Round(100.0 * SkillCastUptimeNoAA, ParserHelper.TimeDigit);
+            //
+            double avgBoons = 0;
+            foreach (long boonDuration in actor.GetBuffPresence(log, start, end).Where(x => log.Buffs.BuffsByIds[x.Key].Classification == BuffClassification.Boon).Select(x => x.Value))
+            {
+                avgBoons += boonDuration;
+            }
+            AvgBoons = Math.Round(avgBoons / duration, ParserHelper.BuffDigit);
+            long activeDuration = actor.GetActiveDuration(log, start, end);
+            AvgActiveBoons = activeDuration > 0 ? Math.Round(avgBoons / activeDuration, ParserHelper.BuffDigit) : 0.0;
+            //
+            double avgCondis = 0;
+            foreach (long conditionDuration in actor.GetBuffPresence(log, start, end).Where(x => log.Buffs.BuffsByIds[x.Key].Classification == BuffClassification.Condition).Select(x => x.Value))
+            {
+                avgCondis += conditionDuration;
+            }
+            AvgConditions = Math.Round(avgCondis / duration, ParserHelper.BuffDigit);
+            AvgActiveConditions = activeDuration > 0 ? Math.Round(avgCondis / activeDuration, ParserHelper.BuffDigit) : 0.0;
+            //
+            if (log.CombatData.HasMovementData && log.FriendlyAgents.Contains(actor.AgentItem) && actor.GetCombatReplayPolledPositions(log).Any(x => x.X > int.MinValue + 1))
+            {
+                StackDist = GetDistanceToTarget(actor, log, start, end, log.StatisticsHelper.GetStackCenterPositions(log));
+                DistToCom = GetDistanceToTarget(actor, log, start, end, log.StatisticsHelper.GetStackCommanderPositions(log));
             }
         }
     }
